# internal/server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Puls",
    "version": "0.1.0"
  },
  "servers": [
    { "url": "http://localhost:8080" }
  ],
  "components": {
    "securitySchemes": {
      "adminAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
      "Device": {
        "type": "object",
        "properties": {
          "id":           { "type": "string" },
          "name":         { "type": "string" },
          "os":           { "type": "string", "enum": ["windows", "linux"] },
          "arch":         { "type": "string" },
          "status":       { "type": "string", "enum": ["online", "offline", "unknown"] },
          "registeredAt": { "type": "string" },
          "lastSeenAt":   { "type": "string" }
        }
      },
      "Heartbeat": {
        "type": "object",
        "properties": {
          "id":            { "type": "integer" },
          "deviceId":      { "type": "string" },
          "receivedAt":    { "type": "string" },
          "cpuPercent":    { "type": "number" },
          "memoryPercent": { "type": "number" },
          "diskPercent":   { "type": "number" },
          "uptimeSeconds": { "type": "integer" },
          "osVersion":     { "type": "string" }
        }
      },
      "DiagnosticResult": {
        "type": "object",
        "properties": {
          "id":          { "type": "integer" },
          "deviceId":    { "type": "string" },
          "requestId":   { "type": "string" },
          "scope":       { "type": "string" },
          "requestedAt": { "type": "string" },
          "receivedAt":  { "type": "string" },
          "payload":     { "type": "object" }
        }
      }
    }
  },
  "paths": {
    "/api/v1/auth/admin-token": {
      "post": {
        "operationId": "getAdminToken",
        "summary": "Issue an admin JWT by presenting the admin secret (PULS_ADMIN_SECRET)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["secret"],
                "properties": {
                  "secret": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "429": { "description": "Rate limit exceeded", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/healthz": {
      "get": {
        "operationId": "healthz",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "status": { "type": "string" } }
                }
              }
            }
          }
        }
      }
    },
    "/openapi.json": {
      "get": {
        "operationId": "getOpenapi",
        "responses": {
          "200": { "content": { "application/json": { "schema": { "type": "object" } } } }
        }
      }
    },
    "/api/v1/devices/register": {
      "post": {
        "operationId": "registerDevice",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "os", "arch", "secret"],
                "properties": {
                  "name":   { "type": "string" },
                  "os":     { "type": "string", "enum": ["windows", "linux"] },
                  "arch":   { "type": "string" },
                  "secret": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deviceId": { "type": "string" },
                    "token":    { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "422": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "429": { "description": "Rate limit exceeded", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/v1/devices": {
      "get": {
        "operationId": "listDevices",
        "security": [{ "adminAuth": [] }],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Device" } }
              }
            }
          },
          "401": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/v1/devices/{id}": {
      "get": {
        "operationId": "getDevice",
        "security": [{ "adminAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/Device" },
                    {
                      "type": "object",
                      "properties": {
                        "recentHeartbeats": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Heartbeat" }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/v1/devices/{id}/diagnose": {
      "post": {
        "operationId": "requestDiagnostics",
        "security": [{ "adminAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "scope": { "type": "string", "enum": ["full", "network", "processes", "storage"] }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DiagnosticResult" }
              }
            }
          },
          "422": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "503": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/v1/devices/{id}/diagnostics": {
      "get": {
        "operationId": "listDiagnostics",
        "security": [{ "adminAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/DiagnosticResult" } }
              }
            }
          }
        }
      }
    },
    "/api/v1/events": {
      "get": {
        "operationId": "streamEvents",
        "summary": "SSE stream of device connect, disconnect, and heartbeat events",
        "security": [{ "adminAuth": [] }],
        "responses": {
          "200": {
            "description": "text/event-stream - each event is `data: {json}\\n\\n`",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type":    { "type": "string", "enum": ["device.connected", "device.disconnected", "device.heartbeat"] },
                    "payload": { "type": "object" }
                  }
                }
              }
            }
          },
          "401": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/v1/ws": {
      "get": {
        "operationId": "connectWebSocket",
        "parameters": [
          { "name": "token", "in": "query", "required": false, "schema": { "type": "string" } }
        ],
        "responses": {
          "101": {},
          "401": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "403": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    }
  }
}